Finalizando AccountsUser e Transfer e fazendo alguma melhorias no cod.

+++ src/controllers/AccountsController.js
@@ -44,9 +44,25 @@ class AccountsController {
       `Você retirou ${value} da conta ${accountNumber}, com sucesso!`
     );
   }
-
   async accountClosure(request, response) {
     const { accountNumber } = request.body;
+    const account = await knex("accounts")
+      .where({ id: accountNumber })
+      .select("balance")
+      .first();
+
+    if (!account) {
+      return response.status(404).json({
+        error: "Conta inválida. Verifique o número da conta e tente novamente.",
+      });
+    }
+
+    if (account.balance > 0) {
+      return response.status(400).json({
+        error:
+          "Conta não pode ser fechada com saldo positivo. Saque seu dinheiro para poder encerrar sua conta.",
+      });
+    }
 
     await knex("accounts").where({ id: accountNumber }).delete();
 
@@ -69,30 +85,38 @@ class AccountsController {
   }
 
   async accountsUser(request, response) {
-    const { accountNumber } = request.params;
+    const { accountNumber } = request.body;
 
-    await knex("accounts").where({ userId: accountNumber });
-
-    // const { accounts } = await knex("accounts")
-    //   .where({ accounts })
-    //   .select("accounts");
+    const accounts = await knex("accounts").where({ userId: accountNumber });
 
     return response.json(accounts);
+  }
+  async transfer(request, response) {
+    const { accountFrom, accountTo, value } = request.body;
 
-    // const [userId, balance] = request.params;
+    const { balance } = await knex("accounts")
+      .where({ id: accountFrom })
+      .select("balance")
+      .first();
 
-    // const accounts = await knex("accounts").where({ userId, balance });
+    if (balance < value) {
+      return response
+        .status(400)
+        .json({ error: "Saldo insuficiente para transferencia." });
+    }
 
-    // return response.json(accounts);
+    await knex("accounts")
+      .where({ id: accountFrom })
+      .decrement({ balance: value });
 
-    // const [user_id] = request.params;
+    await knex("accounts")
+      .where({ id: accountTo })
+      .increment({ balance: value });
 
-    // const accounts = await knex("accounts").where({ user_id });
-
-    // return response.status(201).json(accounts);
+    return response.json(
+      `Você transferiu ${value} para a conta ${accountTo}, com sucesso!`
+    );
   }
 }
 
 module.exports = AccountsController;
-
-// buscar todas as contas dos usuarios
